requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
reportlab
python-dotenv
python-multipart
google-genai
anthropic
groq
sqlalchemy
jinja2
openpyxl
